+++ app/models/user.rb
@@ -10,6 +10,7 @@ class User < ApplicationRecord
   has_many :subscribes, through: :characters
   has_many :worlds, through: :characters
   has_many :static_members, through: :characters
+  has_many :world_fractions, through: :characters
   has_many :identities, dependent: :destroy
   has_one :time_offset, dependent: :destroy